🚚 move elastic-input into atom.io/react-devtools
