+++ packages/atom.io/react-devtools/src/StateEditor.tsx
@@ -3,8 +3,7 @@ import { isJson } from "atom.io/json"
 import { useI, useO } from "atom.io/react"
 import type { FC } from "react"
 
-import { ElasticInput } from "~/packages/hamr/react-elastic-input/src"
-
+import { ElasticInput } from "./elastic-input"
 import { JsonEditor } from "./json-editor"
 
 export const fallback = <T,>(fn: () => T, fallbackValue: T): T => {
