+++ packages/atom.io/react-devtools/src/json-editor/editors-by-type/primitive-editors.tsx
@@ -1,7 +1,6 @@
 import type { ReactElement } from "react"
 
-import { NumberInput, TextInput } from "~/packages/hamr/react-elastic-input/src"
-
+import { NumberInput, TextInput } from "../../elastic-input"
 import type { JsonEditorProps_INTERNAL } from "../json-editor-internal"
 
 export const BooleanEditor = ({
